# docs/swagger.yaml
basePath: /v1
definitions:
  go-boilerplate_internal_dto_article.CreateRequest:
    properties:
      content:
        type: string
      cover_media_id:
        type: integer
      excerpt:
        type: string
      published_at:
        type: string
      slug:
        type: string
      status:
        enum:
        - draft
        - published
        type: string
      title:
        type: string
    required:
    - content
    - cover_media_id
    - excerpt
    - slug
    - title
    type: object
  go-boilerplate_internal_dto_article.ListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/go-boilerplate_internal_dto_article.Response'
        type: array
      meta:
        $ref: '#/definitions/go-boilerplate_pkg_pagination.Meta'
    type: object
  go-boilerplate_internal_dto_article.Response:
    properties:
      content:
        type: string
      cover_media_id:
        type: integer
      created_at:
        type: string
      excerpt:
        type: string
      id:
        type: integer
      published_at:
        type: string
      slug:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      view_count:
        type: integer
    type: object
...
  go-boilerplate_pkg_pagination.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  go-boilerplate_pkg_response.ErrorDetail:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  go-boilerplate_pkg_response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/go-boilerplate_pkg_response.ErrorDetail'
      request_id:
        type: string
      success:
        type: boolean
    type: object
  go-boilerplate_pkg_response.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  go-boilerplate_pkg_response.Response-go-boilerplate_internal_dto_article_ListResponse:
    properties:
      data:
        $ref: '#/definitions/go-boilerplate_internal_dto_article.ListResponse'
      meta:
        $ref: '#/definitions/go-boilerplate_pkg_response.Meta'
      request_id:
        type: string
      success:
        type: boolean
    type: object
  go-boilerplate_pkg_response.Response-go-boilerplate_internal_dto_article_Response:
    properties:
      data:
        $ref: '#/definitions/go-boilerplate_internal_dto_article.Response'
      meta:
        $ref: '#/definitions/go-boilerplate_pkg_response.Meta'
      request_id:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Using a translation service as an example
  title: Go Clean Template API
  version: "1.0"
paths:
  /articles:
    get:
      consumes:
      - application/json
      description: Get a paginated list of articles
      operationId: article-list
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      - description: Article status; omit or leave empty for all statuses
        enum:
        - draft
        - published
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-boilerplate_pkg_response.Response-go-boilerplate_internal_dto_article_ListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go-boilerplate_pkg_response.ErrorResponse'
      summary: List articles
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: Create a new article
      operationId: article-create
      parameters:
      - description: Create Article request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/go-boilerplate_internal_dto_article.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/go-boilerplate_pkg_response.Response-go-boilerplate_internal_dto_article_Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-boilerplate_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go-boilerplate_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create article
      tags:
      - articles
...
swagger: "2.0"